This is a TicTacToe game on Python3! First one of you moves as X, and then the other one moves as O.
I wrote a program that:
Prints an empty field at the beginning of the game.
Creates a game loop where the program asks the user to enter the cell coordinates, analyzes the move for correctness and shows a field with the changes if everything is ok.Ends the game when someone wins or there is a draw.
Outputs the final result after the end of the game.

Good luck gaming!